タッピングタームのデータ型をint8_tからint16_tに変更し、セーブできるように対応 macro_keys.hは使用していなかったので、不要コードの削除と整理。 my_custom_record_keys.hを追加し、保存処理を追加

+++ qmk_firmware/keyboards/keyball/keyball39/keymaps/my_custom/define_variables.h
@@ -23,7 +23,7 @@ typedef union {
     struct {
         int16_t to_clickable_movement;  // クリックレイヤーが有効になるしきい値
 #ifdef DYNAMIC_TAPPING_TERM_ENABLE
-        int8_t tapping_term;           // タッピングタームの設定値
+        int16_t tapping_term;           // タッピングタームの設定値
 #endif
     };
 } user_config_t;
